Overhauled generic methods (for bug fixing) and type-checks (for codegen fusion)
