+++ qpm.json
@@ -4,7 +4,7 @@
   "info": {
     "name": "beatsaber-hook",
     "id": "beatsaber-hook",
-    "version": "0.3.6",
+    "version": "0.3.9",
     "url": "https://github.com/sc2ad/beatsaber-hook",
     "additionalData": {
       "branchName": "so-update",
