--- shared/utils/il2cpp-includes.h
@@ -1,22 +1,0 @@
-#ifndef IL2CPP_INCLUDES_H
-#define IL2CPP_INCLUDES_H
-
-// For including il2cpp properly
-#ifdef _MSC_VER
-#undef _MSC_VER
-#endif
-#ifndef __GNUC__
-#define __GNUC__
-#endif
-
-#define NET_4_0 true
-#include "il2cpp-config.h"
-#include "il2cpp-api-types.h"
-#include "il2cpp-class-internals.h"
-#include "il2cpp-object-internals.h"
-#include "il2cpp-tabledefs.h"
-#include "utils/Il2CppHashMap.h"
-#include "utils/HashUtils.h"
-#include "utils/StringUtils.h"
-
-#endif /* IL2CPP_INCLUDES_H */
